Modified approach to finding areas for cross sections at markers
Before, the approach was to sample the input heat flow image
at the points where contour cross sections were desired. This
failed when those points fell outside or near the border of
the airway interior.

Instead, we create 1000 cross section samples. For each query
point, we find the cross section nearest the sample point and
add it to the output.