Removed the old pipeline files and packed them into the new files. Pipeline.py, Locations.py and Gears.py.
To run the pipeline just call python Pipeline.py. It will analyze the inputs to each pipeline step in order and run each step for the scans that need to be updated.

All of the locations needed for the pipeline are specified in the Locations.py file. Right now they are specific to my computer. Feel free to alter the folders specified in that file but each folder must contain a separate Files and Sha256 folder.